+++ src/render-days.js
@@ -1,0 +1,18 @@
+export default function (data) {
+  const getRepeatedDay = (repeatingDay) => {
+    const dayInLowerCase = repeatingDay.toLowerCase();
+    return (
+      `<input
+        class="visually-hidden card__repeat-day-input"
+        type="checkbox"
+        id="repeat-${dayInLowerCase}-4"
+        name="repeat"
+        value="${dayInLowerCase}"
+        ${data.repeatingDays[repeatingDay] ? `checked` : ``}
+        />
+        <label class="card__repeat-day" for="repeat-${dayInLowerCase}-4">${dayInLowerCase}</label>`
+    );
+  };
+
+  return Object.keys(data.repeatingDays).map(getRepeatedDay).join(``);
+};
